+++ Include/qewindows/shellnodeflags.h
@@ -1,0 +1,1 @@
+#include "../../src/windows/shellnodeflags.h"